refactor: refactor storybook
Поправил структуру, названия файлов, импорты и параметры историй

closes #82
