+++ src/stories/components/addCardButton.stories.ts
@@ -1,26 +1,25 @@
 import type { Meta, StoryObj } from '@storybook/react';
-import { EditCardForm } from '~/entities';
+import { AddCardButton } from '~/features';
 import { withRouter } from 'storybook-addon-react-router-v6';
 
 const meta = {
-  title: 'Components/Edit card form',
-  component: EditCardForm,
+  title: 'Components/Add Card Button',
+  component: AddCardButton,
   decorators: [withRouter],
   parameters: {
     layout: 'centered',
-    width: '320px',
+    width: '360px',
   },
   tags: ['autodocs'],
   argTypes: {},
-} satisfies Meta<typeof EditCardForm>;
+} satisfies Meta<typeof AddCardButton>;
 
 export default meta;
+
 type Story = StoryObj<typeof meta>;
 
-export const Default: Story = {
+export const Primary: Story = {
   args: {
-    isActive: true,
-    cardNumberValue: '1111 1383 0039 3838 49994',
-    barcodeNumberValue: '113839895849854',
+    text: 'Добавить новую карту',
   },
 };
