+++ src/stories/components/addCardForm.stories.ts
@@ -1,19 +1,19 @@
 import type { Meta, StoryObj } from '@storybook/react';
-import { RootLayout } from '~/pages';
+import { AddCardForm } from '~/entities/add-card-form';
 import { withRouter } from 'storybook-addon-react-router-v6';
 
 const meta = {
-  title: 'Pages/RootLayout',
-  component: RootLayout,
+  title: 'Components/Add Card Form',
+  component: AddCardForm,
   decorators: [withRouter],
   parameters: {
     layout: 'centered',
+    width: '360px',
   },
   tags: ['autodocs'],
-} satisfies Meta<typeof RootLayout>;
+} satisfies Meta<typeof AddCardForm>;
 
 export default meta;
-
 type Story = StoryObj<typeof meta>;
 
 export const Default: Story = {};
