+++ src/stories/components/searchChips.stories.ts
@@ -1,17 +1,17 @@
 import type { Meta, StoryObj } from '@storybook/react';
-import { AddCardButton } from '~/features';
+import { SearchChips } from '~/features';
 import { withRouter } from 'storybook-addon-react-router-v6';
+import { chipsLabels } from '~/shared/mock/chips-labels';
 
 const meta = {
-  title: 'Components/AddCardButton',
-  component: AddCardButton,
+  title: 'Components/Search Chips',
+  component: SearchChips,
   decorators: [withRouter],
-  parameters: {
-    layout: 'centered',
-  },
   tags: ['autodocs'],
-  argTypes: {},
-} satisfies Meta<typeof AddCardButton>;
+  parameters: {
+    layout: 'fullscreen',
+  },
+} satisfies Meta<typeof SearchChips>;
 
 export default meta;
 
@@ -19,6 +19,6 @@ type Story = StoryObj<typeof meta>;
 
 export const Primary: Story = {
   args: {
-    text: 'Добавить новую карту',
+    items: chipsLabels,
   },
 };
