+++ src/stories/components/signInForm.stories.ts
@@ -1,22 +1,19 @@
 import type { Meta, StoryObj } from '@storybook/react';
-import { ChipButton } from '~/shared/ui';
+import { SignInForm } from '~/features';
 import { withRouter } from 'storybook-addon-react-router-v6';
 
 const meta = {
-  title: 'UI/ChipButton',
-  component: ChipButton,
+  title: 'Components/Sign In form',
+  component: SignInForm,
   decorators: [withRouter],
   parameters: {
     layout: 'centered',
+    width: '360px',
   },
   tags: ['autodocs'],
-} satisfies Meta<typeof ChipButton>;
+} satisfies Meta<typeof SignInForm>;
 
 export default meta;
 type Story = StoryObj<typeof meta>;
 
-export const Primary: Story = {
-  args: {
-    label: 'category',
-  },
-};
+export const Default: Story = {};
