+++ src/stories/pages/rootLayout.stories.ts
@@ -1,19 +1,19 @@
 import type { Meta, StoryObj } from '@storybook/react';
-import { SignInForm } from '~/features';
+import { RootLayout } from '~/pages';
 import { withRouter } from 'storybook-addon-react-router-v6';
 
 const meta = {
-  title: 'Components/Sign In form',
-  component: SignInForm,
+  title: 'Pages/Root Layout',
+  component: RootLayout,
   decorators: [withRouter],
   parameters: {
-    layout: 'centered',
-    width: '320px',
+    layout: 'fullscreen',
   },
   tags: ['autodocs'],
-} satisfies Meta<typeof SignInForm>;
+} satisfies Meta<typeof RootLayout>;
 
 export default meta;
+
 type Story = StoryObj<typeof meta>;
 
 export const Default: Story = {};
