+++ src/stories/ui/appFooter.stories.ts
@@ -1,16 +1,16 @@
-import { Meta, StoryObj } from '@storybook/react';
+import type { Meta, StoryObj } from '@storybook/react';
 import { AppFooter } from '~/shared/ui';
 
-export default {};
 const meta = {
   title: 'UI/App Footer',
   component: AppFooter,
   parameters: {
-    layout: 'centred',
+    layout: 'fullscreen',
   },
   tags: ['autodocs'],
 } satisfies Meta<typeof AppFooter>;
 
+export default meta;
 type Story = StoryObj<typeof meta>;
 
 export const Default: Story = {};
