+++ src/stories/ui/closeButton.stories.ts
@@ -7,10 +7,8 @@ const meta = {
   component: CloseButton,
   decorators: [withRouter],
   parameters: {
-    // Optional parameter to center the component in the Canvas. More info: https://storybook.js.org/docs/react/configure/story-layout
     layout: 'centered',
   },
-  // This component will have an automatically generated Autodocs entry: https://storybook.js.org/docs/react/writing-docs/autodocs
   tags: ['autodocs'],
 } satisfies Meta<typeof CloseButton>;
