+++ src/stories/widgets/addCardWidget.stories.ts
@@ -1,19 +1,19 @@
-import type { Meta, StoryObj } from '@storybook/react';
-import { AddCardForm } from '~/entities/add-card-form';
+import { Meta, StoryObj } from '@storybook/react';
+import { AddCardWidget } from '~/widgets/add-card';
 import { withRouter } from 'storybook-addon-react-router-v6';
 
 const meta = {
-  title: 'Components/Add card form',
-  component: AddCardForm,
+  title: 'Widgets/Add Card Widget',
+  component: AddCardWidget,
   decorators: [withRouter],
   parameters: {
     layout: 'centered',
-    width: '320px',
+    width: '360px',
   },
   tags: ['autodocs'],
-} satisfies Meta<typeof AddCardForm>;
-
+} satisfies Meta<typeof AddCardWidget>;
 export default meta;
+
 type Story = StoryObj<typeof meta>;
 
 export const Default: Story = {};
