+++ src/widgets/home/index.tsx
@@ -15,8 +15,7 @@ export const Home = () => {
         textAlign="center"
         sx={{ paddingBottom: '1rem', ...paragraphStyle }}
       >
-        Удобный и быстрый доступ к вашим картам лояльности
-        в&nbsp;любом&nbsp;месте
+        Удобный и быстрый доступ к вашим картам лояльности в любом месте
       </Typography>
 
       <Box
